front: merge scenario and stdcm simulation report sheet
