+++ front/src/modules/simulationResult/components/SimulationReportSheet/helpers.ts
