+++ front/src/modules/simulationResult/components/SimulationReportSheet/index.ts
