+++ front/src/modules/simulationResult/components/SimulationReportSheet/style/SimulationReportStyleSheet.ts
